fixed warnings
git-svn-id: file:///srv/svn/repos/haiku/haiku/trunk@27198 a95241bf-73f2-0310-859d-f6bbb57e9c96

+++ src/servers/app/ServerPicture.cpp
@@ -490,7 +490,7 @@ pop_state(View *view)
 	BPoint p(0, 0);
 	view->ConvertToScreenForDrawing(&p);
 	view->Window()->GetDrawingEngine()->SetDrawState(
-		view->CurrentState(), p.x, p.y);
+		view->CurrentState(), (int32)p.x, (int32)p.y);
 }
